Updated views for protection of features for superusers only

+++ products/views.py
@@ -1,11 +1,12 @@
 from django.shortcuts import render, redirect, get_object_or_404, reverse
 from django.contrib import messages
 from django.db.models import Q
+from django.contrib.auth.decorators import login_required
 from django.db.models.functions import Lower
 from .models import Product, Category
 from .forms import ProductForm
 
-# Create your views here.
+# "@login_required" redirects users that arent superusers to the login page.
 
 def all_products(request):
     """ A view to show all products """
@@ -72,10 +73,15 @@ def product_detail(request, product_id):
     return render(request, 'products/product_detail.html', context)
 
 
+@login_required
 def add_product(request):
     """
     Site admins can add a product to the site
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'Only admins can use this feature.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -92,11 +98,14 @@ def add_product(request):
 
     return render(request, template, context)
 
-
+@login_required
 def edit_product(request, product_id):
     """
     Site admins can edit a product to the site
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'Only admins can use this feature.')
+        return redirect(reverse('home'))
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -118,11 +127,14 @@ def edit_product(request, product_id):
 
     return render(request, template, context)
 
-
+@login_required
 def delete_product(request, product_id):
     """
     Site admins can delete a product from the site
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'Only admins can use this feature.')
+        return redirect(reverse('home'))
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product Deleted!')
